+++ src/dbQueries/feeding.ts
@@ -1,9 +1,6 @@
-import { RequestHandler } from "express";
-import { client } from "..";
 import { getCurrentDateTime } from "../helperFunctions";
 import { isCreateFeedingBody } from "../validationFunctions";
-import { PrismaClient } from "@prisma/client";
-import { AuthReqHandler, ReqBuilder } from "../middleware/auth_types";
+import { ReqBuilder } from "../middleware/auth_types";
 
 // Create
 export const createFeeding: ReqBuilder =
