+++ src/lib/constants.ts
@@ -1,0 +1,9 @@
+export const emptyScheduleDays = {
+  monday: false,
+  tuesday: false,
+  wednesday: false,
+  thursday: false,
+  friday: false,
+  saturday: false,
+  sunday: false,
+};
